Move PureParcelEquals to util dir
